+++ frontend/src/components/searchBar/date/Date.tsx
@@ -1,7 +1,7 @@
 import { useState, useRef } from 'react'
 import styled from 'styled-components'
-import { PlaceSection, ModalWrapper } from '../../style/BarStyle'
-import useModalCtrl from '../../customHook/useModalCtrlArray'
+import { PlaceSection, ModalWrapper } from '../../../style/BarStyle'
+import useModalCtrl from '../../../customHook/useModalCtrlArray'
 import ModalCalendar from './ModalCalendar'
 import CheckIn from './CheckIn'
 import CheckOut from './CheckOut'
